Moved .js to src/background, moved fxns from background.js to events.js. Extension only sends headers (issue #166)
