+++ src/background/background.html
@@ -14,12 +14,12 @@ OptMeowt originates here.
 
 
 <!-- Libraries -->
-<script src="libs-js/psl.min.js"></script>
+<!-- <script src="../libs-js/psl.min.js"></script> -->
 
 <!-- Main OptMeowt scripts-->
-<script src="usprivacy.js"></script>
-<script type="module" src="cookie_lists_json.js"></script>
-<script type="module" src="background.js"></script>
+<!-- <script src="./usprivacy.js"></script> -->
+<!-- <script type="module" src="./cookie_lists_json.js"></script> -->
+<script type="module" src="./background.js"></script>
 
 <!-- Compliance analysis mode-->
-<script type="module" src="backgroundanalysis.js"></script>
+<!-- <script type="module" src="./backgroundanalysis.js"></script> -->
